pyproject.toml
[tool.poetry]
name = "postgresql-fastapi-vue-stack-demo"
version = "0.1.0"
description = ""
authors = ["Marek Procházka <m.prochazka2002@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
alembic = "^1.12.0"
sqlmodel = "^0.0.8"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
